refactor: add abstract layer for storage util. Use singleton pattern to avoid duplicating storage object.
