+++ src/components/ScorePage/index.tsx
@@ -1,10 +1,10 @@
 import Layout from "@/components/Common/Layout";
-import { getUserEmail } from "@/utils/storage/auth";
+import { emailStorage } from "@/utils/storage";
 
 export default function ScorePage() {
   return (
     <Layout>
-      <h1>Hello, {getUserEmail() || "user"}! This is Your Score Page</h1>;
+      <h1>Hello, {emailStorage.get() || "user"}! This is Your Score Page</h1>;
     </Layout>
   );
 }
