+++ src/utils/storage/index.ts
@@ -1,22 +1,5 @@
-import { StorageType } from "@/types/utils";
-export function createStorage(type: StorageType = StorageType.LOCAL) {
-  if (typeof window === "undefined") {
-    return;
-  }
-  const storage = type === StorageType.LOCAL ? localStorage : sessionStorage;
-  return {
-    create: function (key: string) {
-      return {
-        set: function (value: string) {
-          storage.setItem(key, value);
-        },
-        get: function () {
-          return storage.getItem(key);
-        },
-        delete: function () {
-          storage.removeItem(key);
-        },
-      };
-    },
-  };
-}
+import createStorage from "@/utils/storage/createStorage";
+
+const storage = createStorage();
+
+export const emailStorage = storage.create("email");
